fix controller cli
